add documentation, add public_transport=stop_position to the checks

+++ src/route_writer.hpp
@@ -23,23 +23,45 @@ enum class RouteType : char {
     SUBWAY
 };
 
-enum class RouteError : size_t {
+/**
+ * All errors this tool can detect on route relations.
+ *
+ * A route has one variable which stores all this errors. Each error is represented by a bit.
+ */
+enum class RouteError : uint32_t {
+    /// route has no errors
     CLEAN = 0,
+    /// Route for railways goes over a non-railway way.
     OVER_NON_RAIL = 1,
+    /// Route for buses goes over a non-highway or non-ferry way.
     OVER_NON_ROAD = 2,
+    /// Route for trolleybuses has no trolley wire.
     NO_TROLLEY_WIRE = 4,
+    /// Route has a gap or is not ordered correctly.
     UNORDERED_GAP = 8,
+    /// Route has a wrong structure in the member list.
     WRONG_STRUCTURE = 16,
+    /// Stops and platforms are missing at the beginning of the route.
     NO_STOPPLTF_AT_FRONT = 32,
+    /// A member which is not a way has an empty role.
     EMPTY_ROLE_NON_WAY = 64,
+    /// A member which is a stop or platform is found after the first highway/ferry/railway member.
     STOPPLTF_AFTER_ROUTE = 128,
+    /// A stop position is not on a way. Not implemented yet.
     STOP_NOT_ON_WAY = 256,
+    /// The relation does not contain any highway/ferry/railway members.
     NO_ROUTE = 512,
+    /// One or many members have an unknown role.
     UNKNOWN_ROLE = 1024,
+    /// The route has an invalid or unknown type (`route=*` tag).
     UNKNOWN_TYPE = 2048,
+    /// A stop member lacks the necessary tags.
     STOP_TAG_MISSING = 4096,
+    /// A platform member lacks the necessary tags.
     PLTF_TAG_MISSING = 8192,
+    /// A stop member is not a node.
     STOP_IS_NOT_NODE = 16384,
+    /// A ferry route uses a way which is not a ferry way.
     NO_FERRY = 32768
 };
 
@@ -51,6 +73,10 @@ inline RouteError operator&(const RouteError&a , const RouteError& b) {
     return static_cast<RouteError>(static_cast<size_t>(a) & static_cast<size_t>(b));
 }
 
+/**
+ * The RouteWriter class writes routes as multilinestrings and their errors (points and linestrings) to
+ * the output dataset.
+ */
 class RouteWriter : public OGROutputBase {
     gdalcpp::Dataset& m_dataset;
     gdalcpp::Layer m_ptv2_routes_valid;
